Install script update

+++ install.sh
@@ -9,16 +9,23 @@ pre_install()
 {
 	if [ -d "$HOME/.wallchanger" ]
 	then
-		# TODO : Save configuration file
 		echo -e "$R Remove existing installation $W"
-		sudo rm -r "$HOME/.wallchanger"
+		sudo rm -r $HOME/.wallchanger
 	fi
 	cd $HOME
 	echo -e "$G Cloning lastest version $W"
 	git clone https://github.com/ziltoidtheomiscient/wallchanger.git .wallchanger
-	cd .wallchanger
-	mv .wallchangerrc ..
-	rm install.sh
+
+	if [ ! -f "$HOME/.wallchangerrc" ]
+	then
+		echo -e "$G No configuration file found, copy the default one $W"
+		mv $HOME/.wallchanger/.wallchangerrc $HOME
+		rm install.sh
+		echo -e "$B Please edit the configuration file : $W"
+		$EDITOR $HOME/.wallchangerrc
+	else
+		echo -e "$G Keeping existing configuration file $W"
+	fi
 }
 
 install_Arch()
@@ -33,12 +40,18 @@ install_Ubuntu()
 	pre_install
 	if [ -f "/etc/init.d/wallchanger" ]
 	then
-		echo -e "$G Remove existing Daemon $W"
-		sudo rm "/etc/init.d/wallchanger"
+		echo -e "$R Remove existing Daemon $W"
+		sudo /etc/init.d/wallchanger stop
+		sudo update-rc.d -f wallchanger remove
+		sudo rm /etc/init.d/wallchanger
 	fi
 	sed -i -e "s/_username_/$USER/g" $HOME/.wallchanger/daemons/ubuntu_daemon
 	sudo cp "$HOME/.wallchanger/daemons/ubuntu_daemon" "/etc/init.d/wallchanger"
-	sudo chmod 0755 "/etc/init.d/wallchanger"
+	sudo chmod 0755 /etc/init.d/wallchanger
+	echo -e "$G Registering wallchanger service $W"
+	sudo update-rc.d wallchanger defaults
+	echo -e "$G Starting the daemon $W"
+	sudo /etc/init.d/wallchanger start
 	post_install
 }
